backend/requirements.txt
fastapi
uvicorn
pydantic
python-dotenv
openai
python-multipart
slowapi
PyJWT
pytest
pytest-cov
pytest-asyncio
anthropic
